+++ CHANGELOG.md
@@ -11,6 +11,7 @@ _Please add a brief description of any changes and any migrations to be performe
 
 
 - [UPDATE] Replaced vue.cli by Vite
+- [NEW] Message can be added to a image or file in the chat
 - [NEW] Added a markdown editor for hub description, house rules, and contact information
 - [NEW] List of the members of a room can be shown now (button in header of the room)